Extract classify_confirm_flow.rs from turn.rs (parent #680)
Third vertical-slice extraction (after anti_pattern_flow, case_record_flow):
move the work-mode / feedback-kind / quality second-pass confirmation
flow out of turn.rs into a new
src/agent/loop_run/classify_confirm_flow.rs sibling module.

Migrated (all pub(super) free fns over &mut Agent):

Entry points:
- classify_with_confirmation (Issue #576 WorkMode wrapper)
- maybe_invoke_work_mode_confirm (Issue #576 gate + dispatch)
- classify_with_feedback_confirm (Issue #579 FeedbackKind wrapper)
- implementation_quality_issue_with_confirm (Issue #580 Quality wrapper)

Private helpers:
- run_work_mode_confirm_attempt
- run_feedback_kind_confirm_attempt
- run_quality_confirm_attempt
- resolve_quality_confirm_issue

No facade re-export (DR3-001). Caller updates:
- turn.rs (x3): self.X(...) → super::classify_confirm_flow::X(self, ...)
- commands.rs (x1): self.classify_with_confirmation(...) →
  super::classify_confirm_flow::classify_with_confirmation(self, ...)
- success.rs (x1): self.classify_with_feedback_confirm(...) →
  super::classify_confirm_flow::classify_with_feedback_confirm(self, ...)

`cargo fix --tests` pruned 8 newly-unused imports from turn.rs's prelude.

turn.rs: 6,557 -> 6,139 (-418 LOC)
Cumulative from 39,489: -33,350 LOC (-84.5%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
